Parse and yield all extracted measurement values as type 'float'
(cherry picked from commit 799d5275a2e174416ae56c7f00dc21e2453ff312)
(cherry picked from commit 26a508d0a5125720b7f7cfbf9e23efbf211cd745)

+++ src/timedelta_isoformat/__init__.py
@@ -31,7 +31,7 @@ class timedelta(datetime.timedelta):
             error_msg = f"{unit} value of {value} exceeds range {bounds}"
             assert value.isdigit(), f"expected a positive integer {unit} component"
             assert value <= limit if inclusive_range else value < limit, error_msg
-            yield unit, int(value)
+            yield unit, float(value)
 
     @staticmethod
     def _fromdatestring(date_string):
